Change behavior for output of solutions to files in pp/clustal format
Overwrite output files for every new solution (during branch and bound search).
In this way, after the search, the best solution is written to the file.
We could change the behavior to write only once after end of search. In both ways,
this enables using the program in place of locarna.